Remove custom error handling for when plugins don't support composer v1
The test was failing because the dependencies used didn't support PHP 8.

However, when I updated them to the oldest version that still supported
PHP 8, then they worked fine against the newest version of composer.

When I dug into it, I realized that because composer 1 was deprecated a
while back, any plugins that don't support the newest version of
composer v1 won't support PHP 8.2, and so the emitted error message is
about deps not working with PHP 8. While still technically possible, the
reality is that the error of allowing a modern PHP version but not a new
version of a deprecated composer major version just isn't going to
happen. And in the worst case that I'm somehow wrong, this isn't that
big of a deal, as the job will fail either way, it just now fails with
a subprocess unknown failure and prints the warning message for a human
to interpret.

So I removed the custom error handling and the associated test.