Avoid unintentional updates from package files
This is an edge case, but if we are using `lockfile-only` mode, the
`pnpm install` may change lockfile `specifiers:` section with something
that won't be present in the final `package.json` (because Dependabot
won't change it). If Dependabot changed `package.json` files, make sure
to rerun `pnpm install --lockfile-only` using the original
`package.json` to amend these unwanted changes.

I'll want to cover this with a spec before I forget.